Code
Possible issues with `fiboMemoization()`.  The count and timing seem not to be changing with increasing `n`, further testing required to determine if that's a bug or actual speed and efficiency due to memoization.  Functions quickly at n>1476, but number is too large to print fully so may display as 'infinity'.

`fiboBruteForce()` seems to function well.  Expected slowness as `n` increases, and could crash a browser n>40.